+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
  * turn on anti-alias by default
  * remove floodfill() method (we won't use it)
  * remove background mask (it's seldom used and will slow down the drawing speed)
+ * set image's default background color to TRANSPARENT
 
 1.0.20
 -----------